[216] Move text preparation into synthefy (#422)
* refactor: move text preparation into synthefy

* test: validate text feature ownership

* fix: preserve released client text compatibility
